Split markdown typography settings
